added screenshots
I could not fit the screenshot in one whole so I did two of them. Also, Sanat and I were both trying to work on part 2 and he said it was fine I had two branches since there was something up with VS code. He said it just mattered that the git ignore worked and the file was hidden.